S08P31A606-215 fix: lng, lat 타입 Double로 변경

+++ kkiri-be/src/main/java/com/lets/kkiri/dto/gps/GpsDto.java
@@ -17,8 +17,8 @@ public class GpsDto {
     private Long id;
     private Long roomId;
     private Long memberId;
-    private int longitude;
-    private int latitude;
+    private Double longitude;
+    private Double latitude;
     @JsonDeserialize(using = LocalDateDeserializer.class)
     private LocalDateTime regDate;
 }
